refactor: layout 제거, GNB 폴더명으로 변경
