+++ apps/web/src/app/[locale]/mypage/page.tsx
@@ -3,8 +3,7 @@ import { Suspense } from 'react';
 import dynamic from 'next/dynamic';
 import { css } from '_panda/css';
 
-import GNB from '@/components/Layout/GNB';
-import Layout from '@/components/Layout/Layout';
+import GNB from '@/components/GNB/GNB';
 
 import RightSection from './RightSection';
 
@@ -12,7 +11,7 @@ const LazyProfileSection = dynamic(() => import('./ProfileSection'), { ssr: fals
 
 function Mypage() {
   return (
-    <Layout>
+    <div className={containerStyle}>
       <GNB />
       <main className={mainStyle}>
         <Suspense fallback={<section></section>}>
@@ -20,11 +19,28 @@ function Mypage() {
         </Suspense>
         <RightSection />
       </main>
-    </Layout>
+    </div>
   );
 }
 
 export default Mypage;
+
+const containerStyle = css({
+  position: 'relative',
+  width: '100%',
+  height: '100%',
+  minHeight: '100vh',
+  minWidth: '1400px',
+  backgroundColor: '#297542',
+  '& > img': {
+    zIndex: -1,
+    objectFit: 'cover',
+  },
+  '& > main': {
+    minHeight: '100vh',
+    zIndex: 1,
+  },
+});
 
 const mainStyle = css({
   padding: '170px 0 0',
