--- apps/web/src/components/Layout/Layout.tsx
@@ -1,28 +1,0 @@
-import type { PropsWithChildren } from 'react';
-import { css } from '_panda/css';
-
-/**
- * @deprecated
- */
-function Layout({ children }: PropsWithChildren) {
-  return <div className={containerStyle}>{children}</div>;
-}
-
-export default Layout;
-
-const containerStyle = css({
-  position: 'relative',
-  width: '100%',
-  height: '100%',
-  minHeight: '100vh',
-  minWidth: '1400px',
-  backgroundColor: '#297542',
-  '& > img': {
-    zIndex: -1,
-    objectFit: 'cover',
-  },
-  '& > main': {
-    minHeight: '100vh',
-    zIndex: 1,
-  },
-});
